created the "comment" route from scratch, components now have a boolean field "commentable" and an array of comment ids, also created the relations between comments and components: if a component is deleted ALL its comments are deleted, if a comment is created its id is stored in its component, if a comment is deleted its id is taken from its components' comment list
added a default page which cannot be deleted, also made sure its url can never be changed

added needed fields to the images, width and height are already calculated and added, now the compressed path and long and lat needs to be implemented

+++ server_side_package/app/controller/pageController.js
@@ -21,18 +21,24 @@ exports.read_a_page     = controller_functions.get_function(page);
 // REMOVING a page entry
 exports.delete_a_page = function(req, res) {
 
-    removeFromComponents(req, res, req.params.id, function(errorOccuredInComponentRemoving) {
-        if(!errorOccuredInComponentRemoving) removeFromLayout(req, res, req.params.id, function(errorOccuredInLayoutRemoving) {
-            if(!errorOccuredInLayoutRemoving) {
-                const page_deletor = controller_functions.delete_function(page);
-                page_deletor(req, res);
-            }
+    // The default page can never be removed
+    if(req.params.id != "Default"){
+        removeFromComponents(req, res, req.params.id, function(errorOccuredInComponentRemoving) {
+            if(!errorOccuredInComponentRemoving) removeFromLayout(req, res, req.params.id, function(errorOccuredInLayoutRemoving) {
+                if(!errorOccuredInLayoutRemoving) {
+                    const page_deletor = controller_functions.delete_function(page);
+                    page_deletor(req, res);
+                }
+            })
         })
-    })
+    } else res.json({ message: "Cannot delete default page!"})
 }
 
 // UPDATING a page entry
 exports.update_a_page = function(req, res) {
+
+    // The default page url can never be changed
+    if(req.params.id == "Default") req.body.url = "/";
 
     addToComponents(req, res, req.params.id, function(errorOccuredInCompAdding) {
         if(!errorOccuredInCompAdding) addToLayout(req, res, req.params.id, function(errorOccuredInLayoutAdding) {
